edit: Users 문서 id provider_email로 수정 및 Users CRUD 관련 ref  모두 수정

+++ src/components/MyPage/EditProfileModal/EditProfileModal.tsx
@@ -65,7 +65,7 @@ const EditProfileModal = ({ setIsModalOpen }: any) => {
     };
 
     setIsModalOpen(false);
-    await updateDoc(doc(db, 'Users', currentUser.userEmail), updateUser);
+    await updateDoc(doc(db, 'Users', `${currentUser.provider}_${currentUser.userEmail}`), updateUser);
     //FIXME: 쿼리 refetch나 invalidateQueries 사용해서 DB 정보로 업데이트 해주는 방법은 없을까?
     setCurrentUser({
       ...currentUser,
@@ -93,7 +93,7 @@ const EditProfileModal = ({ setIsModalOpen }: any) => {
   // [회원탈퇴] 버튼 클릭 시 작동
   const withdrawMembershipHandler = async () => {
     if (confirm('정말 탈퇴하실건가요?🥹🥹🥹🥹')) {
-      await deleteDoc(doc(db, 'Users', currentUser.userEmail));
+      await deleteDoc(doc(db, 'Users', `${currentUser.provider}_${currentUser.userEmail}`));
       alert('회원탈퇴가 완료되었습니다.');
       signOut({ callbackUrl: '/' });
     }
